refactor(ui5-switch): migrate to TS
